Moved navbar to App and display only if logged in
